+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Version 1.0.6
+
+Fixed issues with playlist sounds
+
+Fixed issues with libWrapper conflict message.
+
+Added the option to select multiple by clicking while holding down the shift key.
+
 ## Version 1.0.3
 
 Fixed issues with moving multiple items